Solution to Clear nth Bit of a Number

+++ Clear nth Bit of a Number@@67882a731b4392d8fc1577ec/code.c
@@ -3,7 +3,7 @@ int main(){
     int a,b;
     scanf("%d",&a);
     scanf("%d",&b);
-    int r=((a>>b)&0);
+    int r=(int)((a>>b)&0);
     printf("%d",r);
     return 0;
 }
